requirements.txt
Django==4.0.3
djangorestframework==3.13.1
psycopg2==2.9.3